+++ inst_internal.h
@@ -20,15 +20,15 @@ class BasicBlock {
 /* All the following instructions should never be instantiated explicitly, only by parser */
 class InstBase {
   private:
-    byte opcode;
+    uint16_t opcode;
     opcode_imm_type imm_type;
   public:
-    InstBase (byte opcode) : opcode(opcode) {
+    InstBase (uint16_t opcode) : opcode(opcode) {
       this->imm_type = opcode_table[opcode].imm_type;
     }
     
-    inline byte getOpcode() { return opcode; }
-    inline bool is(byte opc) { return (this->opcode == opc); } 
+    inline uint16_t getOpcode() { return opcode; }
+    inline bool is(uint16_t opc) { return (this->opcode == opc); } 
     inline opcode_imm_type getImmType() { return imm_type; }
 
     virtual void encode_imm (WasmModule &module, bytedeque &bdeq) const;
@@ -42,8 +42,8 @@ class InstBase {
 * */
 class ImmNoneInst : public InstBase {
   public:
-    ImmNoneInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmNoneInst (byte opcode) : 
+    ImmNoneInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmNoneInst (uint16_t opcode) : 
       InstBase(opcode) { };
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -53,8 +53,8 @@ class ImmNoneInst : public InstBase {
 class ImmBlocktInst: public InstBase {
   int64_t type;
   public:
-    ImmBlocktInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmBlocktInst (byte opcode, int64_t type) : 
+    ImmBlocktInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmBlocktInst (uint16_t opcode, int64_t type) : 
       InstBase(opcode), type(type) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -64,8 +64,8 @@ class ImmBlocktInst: public InstBase {
 class ImmLabelInst: public InstBase {
   uint32_t idx;
   public:
-    ImmLabelInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmLabelInst (byte opcode, uint32_t idx) : 
+    ImmLabelInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmLabelInst (uint16_t opcode, uint32_t idx) : 
       InstBase(opcode), idx(idx) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -76,8 +76,8 @@ class ImmLabelsInst: public InstBase {
   std::list<uint32_t> idxs;
   uint32_t def_idx;
   public:
-    ImmLabelsInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmLabelsInst (byte opcode, std::list<uint32_t> idxs, uint32_t def_idx) : 
+    ImmLabelsInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmLabelsInst (uint16_t opcode, std::list<uint32_t> idxs, uint32_t def_idx) : 
       InstBase(opcode), idxs(idxs), def_idx(def_idx) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -87,8 +87,8 @@ class ImmLabelsInst: public InstBase {
 class ImmFuncInst: public InstBase {
   FuncDecl* func;
   public:
-    ImmFuncInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmFuncInst (byte opcode, FuncDecl* func) :
+    ImmFuncInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmFuncInst (uint16_t opcode, FuncDecl* func) :
       InstBase(opcode), func(func) { }
 
     FuncDecl* getFunc() { return this->func; }
@@ -101,8 +101,8 @@ class ImmSigTableInst: public InstBase {
   SigDecl* sig;
   TableDecl* table;
   public:
-    ImmSigTableInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmSigTableInst (byte opcode, SigDecl *sig, TableDecl *table) : 
+    ImmSigTableInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmSigTableInst (uint16_t opcode, SigDecl *sig, TableDecl *table) : 
       InstBase(opcode), sig(sig), table(table) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -112,8 +112,8 @@ class ImmSigTableInst: public InstBase {
 class ImmLocalInst: public InstBase {
   uint32_t idx;
   public:
-    ImmLocalInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmLocalInst (byte opcode, uint32_t idx) : 
+    ImmLocalInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmLocalInst (uint16_t opcode, uint32_t idx) : 
       InstBase(opcode), idx(idx) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -123,8 +123,8 @@ class ImmLocalInst: public InstBase {
 class ImmGlobalInst: public InstBase {
   GlobalDecl* global;
   public:
-    ImmGlobalInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmGlobalInst (byte opcode, GlobalDecl* global) : 
+    ImmGlobalInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmGlobalInst (uint16_t opcode, GlobalDecl* global) : 
       InstBase(opcode), global(global) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -134,8 +134,8 @@ class ImmGlobalInst: public InstBase {
 class ImmTableInst: public InstBase {
   TableDecl* table;
   public:
-    ImmTableInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmTableInst (byte opcode, TableDecl* table) : 
+    ImmTableInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmTableInst (uint16_t opcode, TableDecl* table) : 
       InstBase(opcode), table(table) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -146,8 +146,8 @@ class ImmMemargInst: public InstBase {
   uint32_t align;
   uint32_t offset;
   public:
-    ImmMemargInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmMemargInst (byte opcode, uint32_t align, uint32_t offset) : 
+    ImmMemargInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmMemargInst (uint16_t opcode, uint32_t align, uint32_t offset) : 
       InstBase(opcode), align(align), offset(offset) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -157,8 +157,8 @@ class ImmMemargInst: public InstBase {
 class ImmI32Inst: public InstBase {
   uint32_t value;
   public:
-    ImmI32Inst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmI32Inst (byte opcode, uint32_t value) : 
+    ImmI32Inst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmI32Inst (uint16_t opcode, uint32_t value) : 
       InstBase(opcode), value(value) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -168,8 +168,8 @@ class ImmI32Inst: public InstBase {
 class ImmF64Inst: public InstBase {
   double value;
   public:
-    ImmF64Inst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmF64Inst (byte opcode, double value) : 
+    ImmF64Inst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmF64Inst (uint16_t opcode, double value) : 
       InstBase(opcode), value(value) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -179,8 +179,8 @@ class ImmF64Inst: public InstBase {
 class ImmMemoryInst: public InstBase {
   MemoryDecl* mem;
   public:
-    ImmMemoryInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmMemoryInst (byte opcode, MemoryDecl* mem) : 
+    ImmMemoryInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmMemoryInst (uint16_t opcode, MemoryDecl* mem) : 
       InstBase(opcode), mem(mem) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -191,8 +191,8 @@ class ImmMemoryInst: public InstBase {
 class ImmTagInst: public InstBase {
 
   public:
-    ImmTagInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmTagInst (byte opcode) :
+    ImmTagInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmTagInst (uint16_t opcode) :
       InstBase(opcode) { }
     //void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
 };
@@ -201,8 +201,8 @@ class ImmTagInst: public InstBase {
 class ImmI64Inst: public InstBase {
   int64_t value;  
   public:
-    ImmI64Inst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmI64Inst (byte opcode, double value) : 
+    ImmI64Inst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmI64Inst (uint16_t opcode, double value) : 
       InstBase(opcode), value(value) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -212,8 +212,8 @@ class ImmI64Inst: public InstBase {
 class ImmF32Inst: public InstBase {
   float value;
   public:
-    ImmF32Inst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmF32Inst (byte opcode, float value) : 
+    ImmF32Inst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmF32Inst (uint16_t opcode, float value) : 
       InstBase(opcode), value(value) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -223,8 +223,8 @@ class ImmF32Inst: public InstBase {
 class ImmRefnulltInst: public InstBase {
   wasm_type_t type;
   public:
-    ImmRefnulltInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmRefnulltInst (byte opcode, wasm_type_t type) : 
+    ImmRefnulltInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmRefnulltInst (uint16_t opcode, wasm_type_t type) : 
       InstBase(opcode), type(type) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
@@ -234,8 +234,8 @@ class ImmRefnulltInst: public InstBase {
 class ImmValtsInst: public InstBase {
   std::list<wasm_type_t> types;
   public:
-    ImmValtsInst (WasmModule &module, byte opcode, buffer_t &buf);
-    ImmValtsInst (byte opcode, std::list<wasm_type_t> types) : 
+    ImmValtsInst (WasmModule &module, uint16_t opcode, buffer_t &buf);
+    ImmValtsInst (uint16_t opcode, std::list<wasm_type_t> types) : 
       InstBase(opcode), types(types) { }
 
     void encode_imm (WasmModule &module, bytedeque &bdeq) const override;
