+++ ir.h
@@ -109,7 +109,7 @@ struct GlobalDecl {
 
 struct DataDecl {
   uint32_t flag;
-  byte opcode_offset;
+  uint16_t opcode_offset;
   uint32_t mem_offset;
   bytearr bytes;
 };
@@ -117,7 +117,7 @@ struct DataDecl {
 
 struct ElemDecl {
   uint32_t flag;
-  byte opcode_offset;
+  uint16_t opcode_offset;
   uint32_t table_offset;
   std::list <FuncDecl*> func_indices;
 };
